Extract setting source from getting sources
